Move Subscriptions to com.streamr.client
