+++ client/src/main/java/com/streamr/client/StreamrClient.java
@@ -4,7 +4,6 @@ import com.streamr.client.crypto.Keys;
 import com.streamr.client.crypto.KeysRsa;
 import com.streamr.client.crypto.RsaKeyPair;
 import com.streamr.client.dataunion.DataUnionClient;
-import com.streamr.client.utils.SubscriptionNotFoundException;
 import com.streamr.client.java.util.Objects;
 import com.streamr.client.options.ResendOption;
 import com.streamr.client.options.StreamrClientOptions;
@@ -49,7 +48,6 @@ import com.streamr.client.utils.InMemoryGroupKeyStore;
 import com.streamr.client.utils.KeyExchangeUtil;
 import com.streamr.client.utils.MessageCreationUtil;
 import com.streamr.client.utils.OneTimeResend;
-import com.streamr.client.utils.Subscriptions;
 import java.io.IOException;
 import java.math.BigInteger;
 import java.net.URI;
