--- client/src/main/java/com/streamr/client/utils/AlreadySubscribedException.java
@@ -1,9 +1,0 @@
-package com.streamr.client.utils;
-
-import com.streamr.client.subs.Subscription;
-
-public class AlreadySubscribedException extends RuntimeException {
-    public AlreadySubscribedException(Subscription sub) {
-        super("Already subscribed to streamId: " + sub.getStreamId() + ", partition: " + sub.getPartition());
-    }
-}
