--- client/src/main/java/com/streamr/client/utils/SubscriptionNotFoundException.java
@@ -1,9 +1,0 @@
-package com.streamr.client.utils;
-
-public class SubscriptionNotFoundException extends Exception {
-
-    public SubscriptionNotFoundException(String streamId, int partition) {
-        super("Subscription not found! streamId: " + streamId + ", partition: " + partition);
-    }
-
-}
